# hdms-backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: HDMS API
  description: >
    Hito Device Management System. Source of truth for the HTTP contract —
    see docs/06-api-contract.md. Go server interfaces are generated with
    oapi-codegen; the TypeScript client is generated with openapi-ts.
    Neither is hand-edited.
  version: "0.1.0"
servers:
  - url: /v1
security:
  - adminSession: []
paths:
...
  /devices:
    get:
      operationId: listDevices
      summary: List devices.
      tags: [devices]
      parameters:
        - $ref: "#/components/parameters/DeviceStatusFilter"
        - $ref: "#/components/parameters/CategoryFilter"
        - $ref: "#/components/parameters/QueryFilter"
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceList"
        default:
          $ref: "#/components/responses/ProblemResponse"
    post:
      operationId: createDevice
      summary: Register a new device.
      tags: [devices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceRequest"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /devices/{id}:
    get:
      operationId: getDevice
      summary: Fetch one device.
      tags: [devices]
      parameters:
        - $ref: "#/components/parameters/IDParam"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        default:
          $ref: "#/components/responses/ProblemResponse"
    patch:
      operationId: updateDevice
      summary: Edit a device's fields, including its condition.
      tags: [devices]
      parameters:
        - $ref: "#/components/parameters/IDParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeviceRequest"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /devices/{id}/status:
    post:
      operationId: setDeviceStatus
      summary: Drive the device status lifecycle. Requires a reason for every transition except registration.
      tags: [devices]
      parameters:
        - $ref: "#/components/parameters/IDParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDeviceStatusRequest"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
  /users:
    get:
      operationId: listUsers
      summary: List users.
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/UserStatusFilter"
        - $ref: "#/components/parameters/DepartmentFilter"
        - $ref: "#/components/parameters/QueryFilter"
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
        default:
          $ref: "#/components/responses/ProblemResponse"
    post:
      operationId: createUser
      summary: Register a borrower. Admin-only (FR-45); `registeredBy` is derived from the session, never client-supplied.
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /users/register-with-card:
    post:
      operationId: registerUserWithCard
      summary: Register a borrower and bind a card in one atomic transaction (FR-41, FR-44, FR-59).
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterWithCardRequest"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterWithCardResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /users/{id}:
    get:
      operationId: getUser
      summary: Fetch one user.
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/IDParam"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/ProblemResponse"
    patch:
      operationId: updateUser
      summary: Edit a user's fields. Employee number is not editable — it is the bulk-import matching key.
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/IDParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /users/{id}/suspend:
    post:
      operationId: suspendUser
      summary: Suspend a borrower, recording why.
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/IDParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendUserRequest"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /departments:
    get:
      operationId: listDepartments
      summary: List departments, for the department picker on user forms.
      tags: [users]
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentList"
        default:
          $ref: "#/components/responses/ProblemResponse"

  # --- Credentials --------------------------------------------------------
...
components:
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        extensions:
          type: object
          additionalProperties: true

    # --- Admin auth ---------------------------------------------------------
...
    DeviceStatus:
      type: string
      enum: [available, on_loan, maintenance, retired, lost]
    DeviceCondition:
      type: string
      enum: [good, fair, damaged]
    Device:
      type: object
      required: [id, assetTag, name, categoryId, status, condition, createdAt, updatedAt]
      properties:
        id: { type: string }
        assetTag: { type: string }
        name: { type: string }
        categoryId: { type: string }
        manufacturer: { type: string }
        model: { type: string }
        serialNo: { type: string }
        status:
          $ref: "#/components/schemas/DeviceStatus"
        condition:
          $ref: "#/components/schemas/DeviceCondition"
        homeLocation: { type: string }
        notes: { type: string }
        acquiredOn: { type: string, format: date }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DeviceList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        nextCursor: { type: string }
    CreateDeviceRequest:
      type: object
      required: [assetTag, name, categoryId]
      properties:
        assetTag: { type: string }
        name: { type: string }
        categoryId: { type: string }
        manufacturer: { type: string }
        model: { type: string }
        serialNo: { type: string }
        homeLocation: { type: string }
        notes: { type: string }
        acquiredOn: { type: string, format: date }
    UpdateDeviceRequest:
      type: object
      required: [name, categoryId]
      properties:
        name: { type: string }
        categoryId: { type: string }
        manufacturer: { type: string }
        model: { type: string }
        serialNo: { type: string }
        homeLocation: { type: string }
        notes: { type: string }
        acquiredOn: { type: string, format: date }
        condition:
          $ref: "#/components/schemas/DeviceCondition"
    SetDeviceStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/DeviceStatus"
        reason: { type: string }
...
    UserStatus:
      type: string
      enum: [active, suspended, archived]
    User:
      type: object
      required: [id, employeeNo, fullName, status, registeredAt, registeredBy, updatedAt]
      properties:
        id: { type: string }
        employeeNo: { type: string }
        fullName: { type: string }
        departmentId: { type: string }
        email: { type: string }
        phone: { type: string }
        status:
          $ref: "#/components/schemas/UserStatus"
        notes: { type: string }
        registeredAt: { type: string, format: date-time }
        registeredBy: { type: string }
        updatedAt: { type: string, format: date-time }
    UserList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        nextCursor: { type: string }
    CreateUserRequest:
      type: object
      required: [employeeNo, fullName]
      properties:
        employeeNo: { type: string }
        fullName: { type: string }
        departmentId: { type: string }
        email: { type: string }
        phone: { type: string }
        notes: { type: string }
    UpdateUserRequest:
      type: object
      required: [fullName]
      properties:
        fullName: { type: string }
        departmentId: { type: string }
        email: { type: string }
        phone: { type: string }
        notes: { type: string }
    SuspendUserRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
    RegisterWithCardRequest:
      type: object
      required: [employeeNo, fullName]
      properties:
        employeeNo: { type: string }
        fullName: { type: string }
        departmentId: { type: string }
        email: { type: string }
        phone: { type: string }
        notes: { type: string }
        credentialId:
          type: string
          description: >
            An existing unbound blank credential to bind. If omitted, a
            fresh QR credential is minted for the new user instead.
    RegisterWithCardResponse:
      type: object
      required: [user, credential]
      properties:
        user:
          $ref: "#/components/schemas/User"
        credential:
          $ref: "#/components/schemas/Credential"
        token:
          type: string
          description: >
            The plaintext token, present only when a fresh credential was
            minted (no credentialId was supplied). Absent when an existing
            blank card was bound instead — its token was already disclosed
            at the moment it was printed and is not recoverable now.
    Department:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
    DepartmentList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Department"

    # --- Credentials --------------------------------------------------------
    CredentialKind:
      type: string
      enum: [qr, code128, nfc, rfid, manual]
    SubjectType:
      type: string
      enum: [user, device]
    CredentialStatus:
      type: string
      enum: [active, revoked, lost]
    Credential:
      type: object
      required: [id, subjectType, kind, status, issueSeq, issuedAt, issuedBy, printedCount]
      properties:
        id: { type: string }
        subjectType:
          $ref: "#/components/schemas/SubjectType"
        subjectId:
          type: string
          description: Absent when the credential is unbound blank stock.
        kind:
          $ref: "#/components/schemas/CredentialKind"
        tokenPreview: { type: string }
        label: { type: string }
        status:
          $ref: "#/components/schemas/CredentialStatus"
        issueSeq: { type: integer }
        replacesId: { type: string }
        issuedAt: { type: string, format: date-time }
        issuedBy: { type: string }
        revokedAt: { type: string, format: date-time }
        revokedBy: { type: string }
        revokedReason: { type: string }
        printedCount: { type: integer }
        lastPrintedAt: { type: string, format: date-time }
...
  parameters:
    IDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    CursorParam:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
    QueryFilter:
      name: q
      in: query
      required: false
      schema:
        type: string
    DeviceStatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/DeviceStatus"
    UserStatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/UserStatus"
    CategoryFilter:
      name: category
      in: query
      required: false
      schema:
        type: string
    DepartmentFilter:
      name: department
      in: query
      required: false
      schema:
        type: string
  responses:
    ProblemResponse:
      description: Error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  securitySchemes:
    kioskToken:
      type: http
      scheme: bearer
      description: Long-lived opaque token issued when a kiosk is registered.
    adminSession:
      type: apiKey
      in: cookie
      name: hdms_session
    serviceToken:
      type: http
      scheme: bearer
      description: Used by imports, exports and scheduled jobs.